+++ CHANGES.md
@@ -406,4 +406,6 @@
 ### Pre-release (work-in-progress)
 - Phonak: fix MEDIA and SOUNDNS checking logic with configuration file `config.ini`
 - ReSound/Beltone/Interton/Danavox/Starkey/HIMSA: update archived versions
+- Starkey PatientBase: fix fallback logic
+- Starkey PatientBase: fix verbose output bug
 - Self Update Checker: provide more detail on each release file